+++ src/service.rs
@@ -3,6 +3,8 @@ use crate::yt_dlp;
 use crate::yt_dlp::VideoInfo;
 use dirs;
 use std::process::Command;
+use std::thread::sleep;
+use std::time::Duration;
 
 /// Returns the path of Mpv IPC file.
 /// `mpvy` don't use this directly but giving this as argument to **mpv**
@@ -11,7 +13,7 @@ pub fn ipc_path() -> String {
     format!(
         "{}/mpv/socket",
         dirs::config_dir()
-            .expect("!!! FATAL ERROR !!! Unable to get or find config directory for mpv IPC")
+            .expect("Unexpected Error: Unable to get config directory.")
             .display()
     )
 }
@@ -19,13 +21,14 @@ pub fn ipc_path() -> String {
 /// Spawns `mpv` command with some default arguments to prevent issues.
 /// We are using **local files** because we are want to play audio **downloaded (or cached)**
 /// Even user is offline. Also you can copy your musics to anywhere.
-pub fn mpv(path: &str) -> Result<(), String> {
+fn mpv(path: &str) -> Result<(), String> {
+    info("Service Mpv", "Playing audio with 'mpv'");
     let command = Command::new("mpv")
         .arg("--no-terminal") // Prevent terminal output from mpv
         .arg(format!(
             "--log-file={}/mpvy/log/mpv.log",
             dirs::config_dir()
-                .expect("!!! FATAL ERROR !!! Unable to get or find config directory for mpv log")
+                .expect("Unexpected Error: Unable to get config directory.")
                 .display()
         )) // Change log file path to 'mpvy' log directory
         .arg(format!("--input-ipc-server={}", ipc_path())) // Set up IPC server for controlling mpv
@@ -35,7 +38,7 @@ pub fn mpv(path: &str) -> Result<(), String> {
     if command.is_err() {
         error(
             "Service Mpv",
-            "Unable to play video with mpv. Maybe mpv is not downloaded?",
+            "Unable to play audio with mpv. Maybe mpv is not downloaded?",
         );
         return Err(command.unwrap_err().to_string());
     }
@@ -45,17 +48,14 @@ pub fn mpv(path: &str) -> Result<(), String> {
 /// An function which merges `yt-dlp` APIs for easy usage in coding.
 /// Also uses some optimizations for better user experience.
 /// And logs informative messages
-pub fn play(title: &str) -> Result<yt_dlp::VideoInfo, String> {
-    info(
-        "Service Play",
-        &format!("Requested a play with query '{}'", title),
-    );
+pub fn play(title: &str, duration: u64) -> Result<yt_dlp::VideoInfo, String> {
+    info("Service Play", &format!("Trying to play '{}'.", title));
 
     let video_result: Result<yt_dlp::VideoInfo, String> = yt_dlp::get_info(title); // Get video information (such as duration, title, id)
     let video: VideoInfo = match video_result {
         Ok(info) => info,
         Err(err) => {
-            error("Service Play", "Unable to get video info");
+            error("Service Play", "Unable to get audio info.");
             return Err(err.to_string());
         }
     };
@@ -67,8 +67,15 @@ pub fn play(title: &str) -> Result<yt_dlp::VideoInfo, String> {
         video.id
     );
 
-    // If the video is already downloaded, play it directly
+    // If the audio is already downloaded, play it directly
     if std::fs::exists(&path).unwrap_or(false) {
+        info(
+            "Service Play",
+            "Audio found in the mp3 directory, skipping download.",
+        );
+        let seconds: Duration = Duration::from_secs(duration);
+        // Wait for previous audio to end
+        sleep(seconds);
         let result: Result<(), String> = mpv(&path);
         if result.is_err() {
             return Err(String::from(result.unwrap_err()));
@@ -76,13 +83,14 @@ pub fn play(title: &str) -> Result<yt_dlp::VideoInfo, String> {
         return Ok(video);
     }
 
-    // If the video is not downloaded, download it first
+    // If the audio is not downloaded, download it first
+    info("Service Play", "Downloading audio.");
     let output: Result<(), String> = yt_dlp::download(&video.id);
     if output.is_err() {
-        error("Service Play", "Unable to download video");
+        error("Service Play", "Unable to download audio.");
         return Err(output.unwrap_err().to_string());
     }
-
+    info("Service Play", "Video downloaded successfuly, now playing.");
     let result: Result<(), String> = mpv(&path);
     if result.is_err() {
         return Err(String::from(result.unwrap_err()));